{{cookiecutter.project_directory}}/pyproject.toml
[tool.poetry]
name = "{{cookiecutter.project_directory}}"
version = "0.1.0"
description = ""
authors = ["GGGoingdown <eddiepai0401@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.79.0"
uvicorn = "^0.18.2"
dependency-injector = "^4.39.1"
sentry-sdk = "^1.9.0"
tortoise-orm = "0.19.0"
loguru = "^0.6.0"
aerich = "^0.6.3"
aioredis = "^2.0.1"
tenacity = "^8.0.1"
asyncpg = "0.25.0"

[tool.poetry.dev-dependencies]
pytest = "7.1.1"
pytest-asyncio = "0.18.3"
httpx = "0.22"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
